refactor(master): fix assh to ssh
