+++ pkg/utils/cni/net.go
@@ -22,7 +22,6 @@ import (
 const (
 	defaultInterface = "eth.*|en.*"
 	defaultCIDR      = "100.64.0.0/10"
-	defaultCNIRepo   = "k8s.gcr.io"
 )
 
 type MetaData struct {
