+++ pkg/utils/ssh/connect.go
@@ -1,4 +1,4 @@
-// Copyright © 2021 sealos.
+// Copyright © 2021 Alibaba Group Holding Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,9 +17,15 @@ package ssh
 import (
 	"fmt"
 	"io/ioutil"
+	"net"
 	"os"
+	"path"
+	"path/filepath"
+	"strconv"
 	"strings"
 	"time"
+
+	"github.com/fanux/sealos/pkg/utils/iputils"
 
 	"github.com/fanux/sealos/pkg/utils/logger"
 
@@ -27,78 +33,78 @@ import (
 )
 
 /**
-这里主要是做连接ssh操作的
+  SSH connection operation
 */
-func (ss *SSH) connect(host string) (*ssh.Client, error) {
-	auth := ss.sshAuthMethod(ss.Password, ss.PkFile, ss.PkPassword)
+func (s *SSH) connect(host string) (*ssh.Client, error) {
+	auth := s.sshAuthMethod(s.Password, s.PkFile, s.PkPassword)
 	config := ssh.Config{
 		Ciphers: []string{"aes128-ctr", "aes192-ctr", "aes256-ctr", "aes128-gcm@openssh.com", "arcfour256", "arcfour128", "aes128-cbc", "3des-cbc", "aes192-cbc", "aes256-cbc"},
 	}
-	DefaultTimeout := time.Duration(1) * time.Minute
-	if ss.Timeout == nil {
-		ss.Timeout = &DefaultTimeout
+	DefaultTimeout := time.Duration(15) * time.Second
+	if s.Timeout == nil {
+		s.Timeout = &DefaultTimeout
 	}
 	clientConfig := &ssh.ClientConfig{
-		User:            ss.User,
-		Auth:            auth,
-		Timeout:         *ss.Timeout,
-		Config:          config,
-		HostKeyCallback: ssh.InsecureIgnoreHostKey(),
+		User:    s.User,
+		Auth:    auth,
+		Timeout: *s.Timeout,
+		Config:  config,
+		HostKeyCallback: func(hostname string, remote net.Addr, key ssh.PublicKey) error {
+			return nil
+		},
 	}
-
-	addr := ss.addrReformat(host)
+	ip, port := iputils.GetSSHHostIPAndPort(host)
+	addr := s.addrReformat(ip, port)
 	return ssh.Dial("tcp", addr, clientConfig)
 }
 
-func (ss *SSH) Connect(host string) (*ssh.Session, error) {
-	client, err := ss.connect(host)
+func (s *SSH) Connect(host string) (*ssh.Client, *ssh.Session, error) {
+	client, err := s.connect(host)
 	if err != nil {
-		return nil, err
+		return nil, nil, err
 	}
 
 	session, err := client.NewSession()
 	if err != nil {
-		return nil, err
+		_ = client.Close()
+		return nil, nil, err
 	}
 
 	modes := ssh.TerminalModes{
-		ssh.ECHO:          0,     // disable echoing
+		ssh.ECHO:          0,     //disable echoing
 		ssh.TTY_OP_ISPEED: 14400, // input speed = 14.4kbaud
 		ssh.TTY_OP_OSPEED: 14400, // output speed = 14.4kbaud
 	}
 
 	if err := session.RequestPty("xterm", 80, 40, modes); err != nil {
-		return nil, err
+		_ = session.Close()
+		_ = client.Close()
+		return nil, nil, err
 	}
 
-	return session, nil
+	return client, session, nil
 }
 
-func (ss *SSH) sshAuthMethod(passwd, pkFile, pkPasswd string) (auth []ssh.AuthMethod) {
-	// pkfile存在， 就进行密钥验证， 如果不存在，则跳过密钥验证。
+func (s *SSH) sshAuthMethod(password, pkFile, pkPasswd string) (auth []ssh.AuthMethod) {
 	if fileExist(pkFile) {
-		am, err := ss.sshPrivateKeyMethod(pkFile, pkPasswd)
-		// 获取到密钥验证就添加， 没获取到就直接跳过。
+		am, err := s.sshPrivateKeyMethod(pkFile, pkPasswd)
 		if err == nil {
 			auth = append(auth, am)
 		}
 	}
-	// 密码不为空， 则添加密码验证。
-	if passwd != "" {
-		auth = append(auth, ss.sshPasswordMethod(passwd))
+	if password != "" {
+		auth = append(auth, s.sshPasswordMethod(password))
 	}
-
 	return auth
 }
 
-func fileExist(path string) bool {
-	_, err := os.Stat(path)
-	return err == nil || os.IsExist(err)
-}
+//Authentication with a private key,private key has password and no password to verify in this
+func (s *SSH) sshPrivateKeyMethod(pkFile, pkPassword string) (am ssh.AuthMethod, err error) {
+	pkData, err := ioutil.ReadFile(filepath.Clean(pkFile))
+	if err != nil {
+		return nil, err
+	}
 
-// 使用 pk认证， pk路径为 "$HOME/.ssh/id_rsa", pk有密码和无密码在这里面验证
-func (ss *SSH) sshPrivateKeyMethod(pkFile, pkPassword string) (am ssh.AuthMethod, err error) {
-	pkData := ss.readFile(pkFile)
 	var pk ssh.Signer
 	if pkPassword == "" {
 		pk, err = ssh.ParsePrivateKey(pkData)
@@ -115,24 +121,48 @@ func (ss *SSH) sshPrivateKeyMethod(pkFile, pkPassword string) (am ssh.AuthMethod
 	return ssh.PublicKeys(pk), nil
 }
 
-func (ss *SSH) sshPasswordMethod(passwd string) ssh.AuthMethod {
-	return ssh.Password(passwd)
+func fileExist(path string) bool {
+	_, err := os.Stat(path)
+	return err == nil || os.IsExist(err)
+}
+func (s *SSH) sshPasswordMethod(password string) ssh.AuthMethod {
+	return ssh.Password(password)
 }
 
-// readFile 从文件读取privateKey， 并返回[]byte 无需返回string。
-// 直接返回[]byte， 避免重复 []byte -> string -> []byte
-func (ss *SSH) readFile(name string) []byte {
-	content, err := ioutil.ReadFile(name)
-	if err != nil {
-		logger.Error("[globals] read %s file err is : %s", name, err)
-		os.Exit(1)
-	}
-	return content
-}
-
-func (ss *SSH) addrReformat(host string) string {
+func (s *SSH) addrReformat(host, port string) string {
 	if !strings.Contains(host, ":") {
-		host = fmt.Sprintf("%s:22", host)
+		host = fmt.Sprintf("%s:%s", host, port)
 	}
 	return host
+}
+
+//RemoteFileExist is
+func (s *SSH) IsFileExist(host, remoteFilePath string) bool {
+	// if remote file is
+	// ls -l | grep aa | wc -l
+	remoteFileName := path.Base(remoteFilePath) // aa
+	remoteFileDirName := path.Dir(remoteFilePath)
+	//it's bug: if file is aa.bak, `ls -l | grep aa | wc -l` is 1 ,should use `ll aa 2>/dev/null |wc -l`
+	//remoteFileCommand := fmt.Sprintf("ls -l %s| grep %s | grep -v grep |wc -l", remoteFileDirName, remoteFileName)
+	remoteFileCommand := fmt.Sprintf("ls -l %s/%s 2>/dev/null |wc -l", remoteFileDirName, remoteFileName)
+
+	data, err := s.CmdToString(host, remoteFileCommand, " ")
+	defer func() {
+		if r := recover(); r != nil {
+			logger.Error("[ssh][%s]remoteFileCommand err:%s", host, err)
+		}
+	}()
+	if err != nil {
+		panic(1)
+	}
+	count, err := strconv.Atoi(strings.TrimSpace(data))
+	defer func() {
+		if r := recover(); r != nil {
+			logger.Error("[ssh][%s]RemoteFileExist:%s", host, err)
+		}
+	}()
+	if err != nil {
+		panic(1)
+	}
+	return count != 0
 }
