--- pkg/utils/ssh/hostname.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2021 cuisongliu@qq.com.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package ssh
-
-import (
-	"strings"
-
-	"github.com/fanux/sealos/pkg/utils/logger"
-)
-
-//HostName 判断当前host的hostname
-func HostName(sshConfig SSH, master, host string) string {
-	hostString := sshConfig.CmdToString(master, "kubectl get nodes | grep -v NAME  | awk '{print $1}'", ",")
-	hostName := sshConfig.CmdToString(host, "hostname", "")
-	logger.Debug("hosts %v", hostString)
-	hosts := strings.Split(hostString, ",")
-	var name string
-	for _, h := range hosts {
-		if strings.TrimSpace(h) == "" {
-			continue
-		} else {
-			hh := strings.ToLower(h)
-			fromH := strings.ToLower(hostName)
-			if hh == fromH {
-				name = h
-				break
-			}
-		}
-	}
-	return name
-}
-
-func RemoteHostName(sshConfig SSH, hostIP string) string {
-	hostName := sshConfig.CmdToString(hostIP, "hostname", "")
-	return strings.ToLower(hostName)
-}
